+++ serpentarium/logging.py
@@ -12,7 +12,7 @@ from queue import Queue
 from typing import Collection
 
 
-def configure_child_process_logger(ipc_queue: Queue, level: int = logging.NOTSET):
+def configure_child_process_logger(ipc_logger_queue: Queue, level: int = logging.NOTSET):
     """
     Configures a child process to send all log messages to a queue
 
@@ -21,38 +21,39 @@ def configure_child_process_logger(ipc_queue: Queue, level: int = logging.NOTSET
     Additionally, the only filtering performed is based on log level. By default, all log messages
     are pushed to the queue.
 
-    :param ipc_queue: A Queue object shared between the plugin and the queue servicing process
+    :param ipc_logger_queue: A Queue object shared between the plugin and the queue servicing
+                             process
     :param level: The minimum log level of statements to push to the queue; messages below this log
                   level will be dropped, defaults to logging.NOTSET
     """
     root = logging.getLogger()
-    root.addHandler(QueueHandler(ipc_queue))
+    root.addHandler(QueueHandler(ipc_logger_queue))
     root.setLevel(level)
 
 
 def configure_host_process_logger(
-    ipc_queue: Queue,
+    ipc_logger_queue: Queue,
     handlers: Collection[logging.Handler] = [],
 ) -> QueueListener:
     """
     Configures the root logger to use a QueueListener
 
     A QueueListener can be used to process the log messages from a child process. This function
-    configures a QueueListener to use the provided `ipc_queue` and handlers. It configures the root
-    logger to push log messages from the host process into the `ipc_queue`. Finally, it returns the
-    QueueListener.
+    configures a QueueListener to use the provided `ipc_logger_queue` and handlers. It configures
+    the root logger to push log messages from the host process into the `ipc_logger_queue`. Finally,
+    it returns the QueueListener.
 
     Note that you will need to call `QueueListener.start()`, otherwise the log messages will not be
     processed. See https://docs.python.org/3/library/logging.handlers.html#queuelistener for more
     information about QueueListener
 
-    :param ipc_queue: A Queue shared by the host and child process that stores log messages
+    :param ipc_logger_queue: A Queue shared by the host and child process that stores log messages
     :param handlers: A Collection of LogHandler objects that the QueueListener will use to handle
-                     log messages it pulls off of the ipc_queue
+                     log messages it pulls off of the ipc_logger_queue
 
     :return: An unstarted QueueListener object
     """
     root = logging.getLogger()
-    root.addHandler(logging.handlers.QueueHandler(ipc_queue))
+    root.addHandler(logging.handlers.QueueHandler(ipc_logger_queue))
 
-    return QueueListener(ipc_queue, *handlers, respect_handler_level=True)
+    return QueueListener(ipc_logger_queue, *handlers, respect_handler_level=True)