+++ tests/test_logging.py
@@ -28,33 +28,35 @@ def log_messages(messages_to_log: Iterable[Tuple[int, str]]):
 
 
 def test_child_process_logger__level_notset():
-    spawn_context, ipc_queue, configure_logger_fn = get_logger_config_callback()
+    spawn_context, ipc_logger_queue, configure_logger_fn = get_logger_config_callback()
 
     proc = spawn_context.Process(target=run, args=(configure_logger_fn, LOG_MESSAGES))
     proc.start()
     proc.join(0.15)
 
-    assert_queue_equals(ipc_queue, LOG_MESSAGES)
+    assert_queue_equals(ipc_logger_queue, LOG_MESSAGES)
 
 
 def test_child_process_logger__level_warning():
-    spawn_context, ipc_queue, configure_logger_fn = get_logger_config_callback(logging.WARNING)
+    spawn_context, ipc_logger_queue, configure_logger_fn = get_logger_config_callback(
+        logging.WARNING
+    )
 
     proc = spawn_context.Process(target=run, args=(configure_logger_fn, LOG_MESSAGES))
     proc.start()
     proc.join(0.15)
 
-    assert_queue_equals(ipc_queue, LOG_MESSAGES[2:])
+    assert_queue_equals(ipc_logger_queue, LOG_MESSAGES[2:])
 
 
 def test_configure_queue_listener():
-    ipc_queue = Queue()
+    ipc_logger_queue = Queue()
     test_queue = Queue()
     test_queue_handler = logging.handlers.QueueHandler(test_queue)
     test_queue_handler.setLevel(logging.INFO)
 
     queue_listener = configure_host_process_logger(
-        ipc_queue=ipc_queue, handlers=[test_queue_handler]
+        ipc_logger_queue=ipc_logger_queue, handlers=[test_queue_handler]
     )
 
     try:
@@ -63,5 +65,5 @@ def test_configure_queue_listener():
     finally:
         queue_listener.stop()
 
-    assert ipc_queue.empty()
+    assert ipc_logger_queue.empty()
     assert_queue_equals(test_queue, LOG_MESSAGES[1:])